refactor : 인기 차트 컴포넌트 분리 완료
