+++ src/app/instumentation.ts
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-unresolved */
 /* eslint-disable import/prefer-default-export */
 export async function register() {
   if (process.env.NEXT_RUNTIME === 'nodejs') {
