+++ src/types/chart.ts
@@ -1,0 +1,5 @@
+export interface chart {
+  ranking: number;
+  title: string;
+  thumnailImage: string;
+}
